==> jupyter_notebooks/06_ModelAndEvaluate.ipynb <==
# CI Portfolio Project 5 - Filter Maintenance Predictor 2022
## **ML Model - Predict Optimal Filter Change (OFC)**

## Objectives

Answer [Business Requirement 2](https://github.com/roeszler/filter-maintenance-predictor/blob/main/README.md#business-requirements) :
*   Fit and evaluate a **regression model** to predict the optimal interval to change a replaceable part
*   Fit and evaluate a **classification model** to predict the optimal interval to change a replaceable part should the regressor not perform well.

## Inputs

Data cleaning and feature engineering from their respective notebooks:
* inputs/datasets/cleaned/df_train.csv
* inputs/datasets/cleaned/df_test.csv
* inputs/datasets/cleaned/df_validate.csv

## Outputs

* Train set (features and target)
* Test set (features and target)
* Validation set (features and target)
* ML pipeline to predict Optimal Filter Change
* A map of the labels
* Feature Importance Plot



---
